Move helper functions to common file
